+++ pom.xml
@@ -23,7 +23,7 @@
     <groupId>com.forgerock.securebanking.uk</groupId>
     <artifactId>securebanking-openbanking-uk</artifactId>
     <packaging>pom</packaging>
-    <version>1.1.3-SNAPSHOT</version>
+    <version>1.1.4-SNAPSHOT</version>
     <name>Secure Banking access Toolkit library for Open Banking UK specs</name>
     <url>https://github.com/SecureBankingAccessToolkit/securebanking-openbanking-uk</url>
