+++ securebanking-openbanking-uk-forgerock-datamodel/pom.xml
@@ -30,7 +30,7 @@
     <parent>
         <groupId>com.forgerock.securebanking.uk</groupId>
         <artifactId>securebanking-openbanking-uk</artifactId>
-        <version>1.1.3-SNAPSHOT</version>
+        <version>1.1.4-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
